+++ client/app/CatalogList.tsx
@@ -1,16 +1,21 @@
 import { ScrollView, TextInput, View } from 'react-native';
 import CatalogItem from '@/components/CatalogItem';
 
+export {
+  // Catch any errors thrown by the Layout component.
+  ErrorBoundary,
+} from 'expo-router';
+
 const CatalogList = () => {
   return (
     <ScrollView>
       <View>
-      <TextInput
+        <TextInput
           style={{
-            borderColor: "#ddd",
+            borderColor: '#ddd',
             borderWidth: 1,
             padding: 10,
-            paddingLeft:20,
+            paddingLeft: 20,
             borderRadius: 30,
             marginTop: 8,
             marginBottom: 8,
@@ -18,22 +23,22 @@ const CatalogList = () => {
           placeholder="search"
         />
         <CatalogItem
-          title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          title="3X2"
+          img={require('../assets/images/WindowPreview.png')}
           size="100cm x 200cm"
           material={'frame material'}
           feature={'feature'}
         />
         <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
         />
         <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
@@ -41,29 +46,29 @@ const CatalogList = () => {
 
         <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
         />
-         <CatalogItem
+        <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
         />
-         <CatalogItem
+        <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
         />
 
-<CatalogItem
+        <CatalogItem
           title="Hello"
-          img="https://reactnative.dev/img/tiny_logo.png"
+          img={require('../assets/images/WindowPreview.png')}
           size="Frame size"
           material={'frame material'}
           feature={'feature'}
