+++ client/components/CatalogItem.tsx
@@ -1,10 +1,18 @@
 import { Link } from 'expo-router';
 import { FC } from 'react';
-import { Image, ImageSourcePropType, Pressable, Text, View } from 'react-native';
+import {
+  Image,
+  ImageSourcePropType,
+  Pressable,
+  StyleSheet,
+  Text,
+  View,
+} from 'react-native';
+import { AntDesign } from '@expo/vector-icons';
 
 type CatalogItemProps = {
   title: string;
-  img: string;
+  img: ImageSourcePropType;
   size: string;
   material: string;
   feature: string;
@@ -18,24 +26,64 @@ const CatalogItem: FC<CatalogItemProps> = ({
   feature,
 }) => {
   return (
-    <Link href="/GridConfigurator"  asChild>
+    <Link href="/Camera" asChild>
       <Pressable>
-      <View
-      style={{ flexDirection: 'row', marginBottom: 10, marginTop: 10, gap: 10, borderWidth:1, borderColor:"#ccc", padding:8 }}
-    >
-      <Image style={{ width: 100, height: 100 }} source={{uri: img}} />
-      <View style={{ flexDirection: 'column', gap: 5 }}>
-        <Text style={{ fontSize: 20, fontWeight: 'bold' }}>{title}</Text>
-        <Text>{size}</Text>
-        <Text>{material}</Text>
-        <Text>{feature}</Text>
-      </View>
-    </View>
+        <View
+          style={{
+            flexDirection: 'row',
+            marginBottom: 10,
+            marginTop: 10,
+            gap: 10,
+            borderWidth: 1,
+            borderColor: '#ccc',
+            padding: 8,
+          }}
+        >
+          <Image style={{ width: '50%' }} source={img} />
+          <View style={{ flexDirection: 'column', gap: 5 }}>
+            {/* Glanzings code start here */}
+            <View style={{ display: 'flex', gap: 5, flexDirection: 'row' }}>
+              <View style={{ ...styles.button, backgroundColor: '#ccc' }}>
+                <Image
+                  style={{ width: '50%', height: '80%' }}
+                  source={require('../assets/images/Glazing.png')}
+                />
+              </View>
+              <View style={{ ...styles.button, borderWidth: 1 }}>
+                <Text>2x</Text>
+              </View>
+            </View>
+            {/* Glanzings code End here */}
+
+            <Text style={{ fontSize: 20, fontWeight: 'bold' }}>{title}</Text>
+            <Text>{size}</Text>
+            <Text>{material}</Text>
+            <Text>{feature}</Text>
+            <AntDesign
+              style={{ position: 'absolute', right: -40, bottom: 30 }}
+              name="right"
+              size={30}
+              color="#ccc"
+            />
+          </View>
+        </View>
       </Pressable>
-     
     </Link>
-    
   );
 };
-
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  button: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    width: 60,
+    height: 60,
+    borderRadius: 50,
+  },
+});
 export default CatalogItem;
